[CNEUR-116]: Removing dead code, changing IvocVect datastructure with simple and nice template class implementation.
Ran the code with Coverity, fixed potential memory violations and
identified couple of places which might have uninitialized variables.


CoreNEURON Repo SHA: BlueBrain/CoreNeuron@cd01b52a7e5dd518678575a2e5b4c08dd5b27167

+++ src/coreneuron/nrniv/vrecitem.h
@@ -73,7 +73,7 @@ public:
 	IvocVect* y_;
 	IvocVect* t_;
 	double dt_;
-	int current_index_;
+    size_t current_index_;
 
 	PlayRecordEvent* e_;
 };
@@ -98,9 +98,9 @@ public:
 	IvocVect* y_;
 	IvocVect* t_;
 	IvocVect* discon_indices_;
-	int last_index_;
-	int discon_index_;
-	int ubound_index_;
+    size_t last_index_;
+    size_t discon_index_;
+    size_t ubound_index_;
 
 	PlayRecordEvent* e_;
 };
